Shawn's physics contribution and other stuff
Added sou nds to locked doors.

Added Gravity use and Animation to falling rocks trap as well as refined
the trap a bit. Then added sound. Overall should be a good contribution
to assignment 8. Am thinking of adding something similar to when the
light button is pressed.

Applied giovanni's rat code to the dumb monster to see how it would
behave. Still working on my own AI code... having trouble with adjacency
lists and matrixes right now separately and not in commit. added the
audio to the assets folder and stuff and put it on our trello.