added the finale rocks falling when the win button pressed
dimmed the lights again. For some reason the lights I have set up to
turn on during the 5 seconds after pressing the win button doesn't t urn
on.... I'm a bit too tired to fix it right now. will do it when I wake
up. But I would appreciate it if someone else also took a look at it for
additional opinions on how to fix.