created a simple AI using older techniques
It is supposed to make the models go towards the selected door.
Unfortunately I can't seem to make the models go into the right
direction. They don't seem to be bounded to the X axis even if I put on
the freeze constraints. I think there is something off about their
initial orientation. Does anyone know how to change that?