added movement to the monster in Alpha
Ideally I wanted it to go down the stairs as well to follow after its
prey.

A bug I found but don't have the time to fix at the moment (anyone can
fix it) is that touching the monster here doesn't lead to a game over,
it instead leads to level.end for some reason.